rename description_plot file
- rename file to description_variable
- rename also test and pandas implementation of file
