+++ src/pandas_profiling/model/description_variable.py
